+++ qiskit/backends/ibmq/ibmqprovider.py
@@ -11,6 +11,7 @@ from IBMQuantumExperience import IBMQuantumExperience
 from qiskit._util import _camel_case_to_snake_case
 from qiskit.backends.baseprovider import BaseProvider
 from qiskit.backends.ibmq.ibmqbackend import IBMQBackend
+from qiskit._util import _parse_ibmq_credentials
 
 
 class IBMQProvider(BaseProvider):
@@ -19,9 +20,10 @@ class IBMQProvider(BaseProvider):
                  hub=None, group=None, project=None, proxies=None, verify=True):
         super().__init__()
 
+        url = _parse_ibmq_credentials(url, hub, group, project)
+
         # Get a connection to IBMQuantumExperience.
-        self._api = self._authenticate(token, url,
-                                       hub, group, project, proxies, verify)
+        self._api = self._authenticate(token, url, proxies=proxies, verify=verify)
 
         # Populate the list of remote backends.
         self.backends = self._discover_remote_backends()
@@ -29,9 +31,6 @@ class IBMQProvider(BaseProvider):
         # authentication attributes, which uniquely identify the provider instance
         self._token = token
         self._url = url
-        self._hub = hub
-        self._group = group
-        self._project = project
         self._proxies = proxies
         self._verify = verify
 
@@ -67,9 +66,7 @@ class IBMQProvider(BaseProvider):
             }
 
     @classmethod
-    def _authenticate(cls, token, url,
-                      hub=None, group=None, project=None, proxies=None,
-                      verify=True):
+    def _authenticate(cls, token, url, proxies=None, verify=True):
         """
         Authenticate against the IBMQuantumExperience API.
 
@@ -83,13 +80,6 @@ class IBMQProvider(BaseProvider):
             config_dict = {
                 'url': url,
             }
-            # Only append hub/group/project if they are different than None.
-            if all([hub, group, project]):
-                config_dict.update({
-                    'hub': hub,
-                    'group': group,
-                    'project': project
-                })
             if proxies:
                 config_dict['proxies'] = proxies
             return IBMQuantumExperience(token, config_dict, verify)
@@ -144,9 +134,7 @@ class IBMQProvider(BaseProvider):
 
     def __eq__(self, other):
         try:
-            equality = (self._token == other._token and self._url == other._url and
-                        self._hub == other._hub and self._group == other._group and
-                        self._project == other._project)
+            equality = (self._token == other._token and self._url == other._url)
         except AttributeError:
             equality = False
         return equality